requirements.txt
PyQt6>=6.4.0
opencv-python>=4.8.0
numpy<2.0.0
Pillow>=10.0.0
pywin32>=306
psutil>=5.9.0
torch>=2.0.0 